src/requirements.txt
Django==1.8.4
Pillow==3.0.0
argparse==1.2.1
beautifulsoup4==4.4.1
boto==2.38.0
boto3==1.4.0
botocore==1.4.43
cobble==0.1.1
django-crispy-forms==1.5.2
django-picklefield==0.3.2
django-registration-redux==1.2
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
django-storages==1.1.8
docx2html==0.2.3
docx2txt==0.4
geopy==1.11.0
lxml==3.5.0
mammoth==0.3.25
parsimonious==0.5
python-dateutil==2.4.2
python-docx==0.8.5
rauth==0.7.2
requests==2.8.1
six==1.10.0
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.1.1
psycopg2==2.5.1
static==0.4